more added to taggable app

+++ apps/taggable/models.py
@@ -5,30 +5,56 @@ class Taggable(db.Model):
     tags = db.StringListProperty()
 
     @classmethod
+    def get_taggable(cls, obj):
+        try: 
+           return (obj.taggable_set)[0]
+        except:
+            return None
+    
+    @classmethod
     def get_by_tag(cls, tag):
        return Taggable.all().filter('tags =', tag) 
-
+   
+    @classmethod
+    def clear_tags(cls, obj):
+        '''clear_tags: remove all tag from an object'''
+        taggable = Taggable.get_taggable(obj)
+        if taggable is not None: 
+            taggable.delete()
+            
+    @classmethod 
+    def set_tags(cls, obj, tags):
+        taggable = Taggable.get_taggable(obj)
+        if taggable is None:
+            taggable = Taggable(target = obj, tags = tags)  
+        else:
+            taggable.tags = tags
+        taggable.put()
+        
     @classmethod
     def add_tags(cls, obj, tags):
-       try: 
-           taggable = (obj.taggable_set)[0]
-           taggable.tags = tags
-       except: 
-           taggable = Taggable(target = obj, tags = tags)
-       taggable.put()
+        taggable = Taggable.get_taggable(obj)
+        if taggable is None: 
+          taggable = Taggable(target = obj, tags = tags)           
+        else: 
+            for tag in tags:
+               if tag not in taggable.tags:
+                   taggable.tags.append(tag)
+        taggable.put()
        
     @classmethod
     def remove_tags(cls, obj, tags):
-        try:
-            taggable = (obj.taggable_set)[0]
-            taggable.tags = []
+        taggable = Taggable.get_taggable(obj)
+        if taggable is not None: 
+            for tag in tags: 
+                if tag in taggable.tags: 
+                    taggable.tags.remove(tag)
             
             if len(taggable.tags) == 0 : 
                 taggable.delete()
             else:
                 taggable.put()
-        except:
-            pass
+
    
 class Tag(db.Model):
     name = db.StringProperty(required = True)
